// src/api/generated/models/user.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * ToggleFlow API
 * ToggleFlow is a Feature Flag Management Platform that allows teams to
 * manage feature flags, environments, experiments, API keys, segments,
 * webhooks, and analytics through a REST API.
 *
 * OpenAPI spec version: 1.0.0
 */

export interface User {
  id?: string;
  email?: string;
  username?: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  avatar?: string | null;
  isEmailVerified?: boolean;
  isActive?: boolean;
  createdAt?: string;
  updatedAt?: string;
}